n6-orin: de-ROS to standalone CMake + port TensorRT 8.6->10, build on Jetson Orin
Make AirSLAM build and run as a standalone (non-ROS) SLAM backend on the N6 Jetson
Orin (Ubuntu 24.04, TensorRT 10.16 / CUDA 13.2), validated on TUM-VI room1
stereo+IMU (2821 frames, 685 keyframes, ~90k map points, metric trajectory).

De-ROS (no ROS Noetic on 24.04):
- ros_publisher -> no-op (rviz live-viz dropped; trajectories still saved directly);
  drop ROS/cv_bridge/msgs includes, keep the message structs + public API.
- add include/ros_shim.h: header-only stand-ins for ros::NodeHandle/init/ok/
  spinOnce/shutdown/Time/Rate so the core needs no ROS.
- demo mains: ros::param -> positional argv. Replace boost::bind thread spawns with
  native std::thread member-function form (boost::bind was only reachable via the
  now-removed thread_publisher include).
- CMakeLists: drop catkin + gflags/glog (all unused); standalone lib + 4 demos.

TensorRT 8.6 -> 10 port:
- buffers.h BufferManager: binding-index API -> named-tensor API (getNbIOTensors/
  getIOTensorName/getTensorShape/getTensorDataType/getTensorIOMode) + setTensorAddresses().
- super_point/plnet/super_glue/light_glue: createNetworkV2(0), setInputShape/
  getTensorShape by name, enqueueV3 + stream sync (executeV2 removed).
- tensorrtbuffer: build only logger.cpp (sample_options/sample_utils are trtexec CLI
  infra using removed TRT8 APIs); drop Caffe/UFF parser entrypoints (removed in TRT10);
  fix roundUp() for int64 Dims::d.

Build deps:
- cmake/FindG2O.cmake: drop unfound optional g2o libs from G2O_LIBRARIES (guard with
  DEFINED, since if(var) is false when the joined value ends in -NOTFOUND).
- define G2O_USE_VENDORED_CERES (our g2o uses vendored Ceres; FindG2O doesn't propagate it).
- line_processor.h: drop unused opencv2/ximgproc.hpp (contrib; JetPack OpenCV lacks it).

Models / configs:
- tools/fix_onnx_int_types_for_trt10.py + apply to superpoint_v1_sim_int32.onnx: the
  int32-typecast model mixes Int32/Int64 into Concat/elementwise ops, which TRT10
  rejects; insert int32->int64 casts to unify (TRT10 supports int64 natively).
- configs/camera/tumvi.yaml + configs/visual_odometry/vo_tumvi.yaml for validation.

+++ 3rdparty/tensorrtbuffer/include/safe_common.h
@@ -110,7 +110,9 @@ namespace tensorrt_buffer
     {
         if (vecDim >= 0)
         {
-            dims.d[vecDim] = roundUp(dims.d[vecDim], comps);
+            // TensorRT 10: Dims::d is int64_t (was int32_t in TRT 8), so roundUp's
+            // template can't deduce a single T from (int64_t, int32_t). Widen comps.
+            dims.d[vecDim] = roundUp(dims.d[vecDim], static_cast<int64_t>(comps));
         }
         return tensorrt_buffer::volume(dims) * std::max(batch, 1);
     }
